rtcpu: debug: Pass test memory physical address
When using debug memory to emulate hardware sync points,
they are accessed with physical streamID, in which case
a physical address has to be used instead of a iova.

To enable this use case, added keeping track of
physical address for test memories and passing that
address down to RCE with the memory test request.

The physical address is only valid for first page,
so also page size is passed down to RCE.

Bug 200642641

Signed-off-by: Kalle Jokiniemi <kjokiniemi@nvidia.com>
Change-Id: I35e3f6e5a3262289be814318b7e1b8dfb9444725
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2467874
Reviewed-by: Mika Liljeberg <mliljeberg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit