video: tegra: Fix build issue to make NvMap as OOT
Fix following build issues while making NvMap as OOT loadable module.
1)Add new DMA_ATTR* for NvMap module in case of loadable module to fix
build issues.
2)Making NvMap as OOT module results compilation failure as definition
of CONFIG_NVMAP_FD_START is not present in the nvmap, as it is a
kconfig. Hence in case of kstable, define this macro locally in nvmap
with the value same as default value from the Kconfig.
3)CONFIG_NVMAP_PAGE_POOLS is present in the Kconfig and not in the
nvmap. It's default value in Kconfig is y. Fix this issue by always
enabling it, in case of kstable.

Bug 2879962

Change-Id: Ibdb6d8a3547a8ab23a612bf4bcb1ffeb67de55fb
Signed-off-by: Puneet Saxena <puneets@nvidia.com>
Signed-off-by: ketanp <ketanp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2472429
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>