DOWNSTREAM: iommu: arm-smmu: io-pgtable: Add dma_sync API
Add an API to iommu's and pagetables that is similar to the TLB flushing
API, but it instead flushes the memory out to the device.
This is called from the iommu framework because the smmu framework isn't
aware of the whole buffer.
Under the old flushing, if the iommu framework's worst case scenario would
be a buffer mapping to 254 contiguous PTE entries. The old framework would
do 254 dma_sync calls, the new framework syncs all of the contiguous page
table entries at once.
Enable the coherent_walk now that we have the dma_sync API enabled so we no
longer do flushing in the normal io-pgtable map/unmaps.

Bug 2322310
Bug 200633784

Change-Id: I3f56c3ee2a76ce4bbb5d5edc14e09cc593473c17
Signed-off-by: David Gilhooley <dgilhooley@nvidia.com>
(cherry picked from commit 0033244b281683a87d62508181a550768adcb43f)
Reviewed-on: https://git-master.nvidia.com/r/1796278
Signed-off-by: Ashish Mhetre <amhetre@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2489383
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>