ASoC: tegra-alt: Remove AUD_MCLK reset
T186 has AUD_MCLK reset but this is unused. This essentially means
programming this has no effect and it is safe for SW to ignore this.
Thus this reset is cleaned up from drivers and DT.

Bug 200696954

Note: At this point there is no ap_audio coverage for T186/T194 boards
in GVS for kernel-4.14 and the coverage is already moved to kernel-5.10.
However ap_camera is still running on kernel-4.14 and is flagging some
warnings in GVS due to sound card registration failure, when AUD_MCLK
reset is removed from the DT. See below for more details:
  http://vrl-preprod.nvidia.com/showjob.php?job=411115151

To unblock above driver is updated.

Change-Id: Ibcc535aa3d22a6c289ac95c1c2ca1c1def260035
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2491166
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>