dts: removing clocks which is not needed.
Some Clocks ids were added before as it was suspected to get in use for
programming but now seems they donot needed for programming.
So, whichever is sure that may not be needed for programming, removing
from the dts.

Editing names of few clocks to match with naming convention of other
clocks, else it will look confusing.

Jira Id TDS-6940

Change-Id: I8aaf138eadd127514c17348ea8d9c8b34360b3d5
Signed-off-by: Gaurav Madheshia <gmadheshia@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/soc/t23x/+/2499440
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-by: Arun Swain <arswain@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit