PCI: tegra: Add support to read unified DT nodes
PCIe device-tree node entries are unified across different kernel
versions. Because of that, it is required to add support to look for
different names of the same entry before bailing out as an error when it
is not found in the first instance of the search.

Bug 3088099

Change-Id: I35fd02650aa23d8b702980d79a24d3faea5e4e02
Signed-off-by: Vidya Sagar <vidyas@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2503625
Reviewed-by: Manikanta Maddireddy <mmaddireddy@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit