video: tegra: host: Update Nvhost/Vhost module IDs to be sequential
Current Nvhost module IDs and Vhost module IDs are not sequential since
bits[23:16] are used as instance ID of the module on the ID value.
But this can make wrong result on range check of the ID value. So this
change updates Nvhost module IDs and Vhost module IDs to be sequential.

Jira HOSTXV-1595

Change-Id: I301a5a97650c13f5256aec70de9d32ac3b02ea7f
Signed-off-by: Jinyoung Park <jinyoungp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2504497
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit