video: tegra: host: Do not add waiter for expired fence
If a userspace application tried to create a sync_file object for
syncpoint ID 0, the interrupt threshold for ID 0 got overridden with
the specified threshold. ID 0 has special handling in the interrupt
routine to detect increments with the assumption that the threshold
is always 1. This caused an interrupt flood. Fix by not creating
waiters for fences that are already expired.

Bug 200705261

Change-Id: I0ce54c696d938a7b1436af1a9866252801884e7a
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2504539
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>