mmc: host: Enable VQMMC regulator if valid
The host sdhci layer maintains a flag "enable_vqmmc" to
enable the vqmmc regulator which powers the SDMMC controller.
"enable_vqmmc" flag is set if the regulator is parsed only inside
sdhci_setup_host.
Tegra sdhci driver parses the regulator nodes in the probe itself due to
dependency on the PMIC driver.
Set "enable_vqmmc" if a valid regulator is populated for the vqmmc field,
irrespective of the function in which it is parsed.

Bug 200684231
Bug 200711968
Bug 200706008

Change-Id: I6356be1ae953a14a2f1bc33c3de88445ce21fcaa
Signed-off-by: Aniruddha Tvs Rao <anrao@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2504720
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit