ASoC: audio-graph: Use of_node and DAI for normal DAI link names
When there are multiple instances of components, conflicts are seen
with DAI link names as it depends on CPU and Codec DAI names alone.
To avoid this use DT device node reference along with DAI name for
the DAI link name.

This change is inspired from commit fa4d1312218b ("ASoC: audio-graph:
Use of_node and DAI for DPCM DAI link names") which was done for
DPCM DAI links.

Bug 200692799

Change-Id: I80083974e910f8fa6e3b29a6753f647b1341cd08
Signed-off-by: Sameer Pujar <spujar@nvidia.com>