video: tegra: nvmap: Remove _initcall
In order to make NvMap as a loadable kernel module, remove _initcalls
used i.e. core_initcall and fs_initcall.
kstable does not allow change in the DT, hence instead of making any
change in the DT, skip the carveout initialization part for the
kstable. We need some cleaner way in future to achieve this.

Bug 2879962

Change-Id: I48efa17342edc46cbf7afcdf166435abe123386c
Signed-off-by: Ketan Patil <ketanp@nvidia.com>
Signed-off-by: Puneet Saxena <puneets@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2383559
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit