video: tegra: nvmap: Correct vmap implementation
The definition of dma_buf_vmap has been changed in kernel > 5.10 The
vmap operation on dmabuf expects int as a return type, earlier it's
return type was void *. It expects 0 on success and a negative error
code otherwise. As we implement the vmap operation in nvmap driver,
hence it should return 0 if the mapping is successful or negative
error code on failure. The kernel has also introduced the new struct
dma_buf_map for storing the address and memory type after mapping.
Hence store the mapped address into this struct, instead of directly
returning it.

Bug 2879962

Change-Id: I7e19b093a7a351fb0737aa1c28f39218aa3a9df4
Signed-off-by: ketanp <ketanp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2511968
Reviewed-by: Pritesh Raithatha <praithatha@nvidia.com>
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit