thermal: userspace_alert: Fix automatic loading of userspace_alert
The userspace_alert driver is not automatically loaded on boot because
the MODULE_DEVICE_TABLE definition for the driver is missing. Add the
MODULE_DEVICE_TABLE for the userspace_alert driver so that the modalias
is created and the driver is automatically loaded on boot.

JIRA LS-32
Bug 200721211

Change-Id: I2cca5bd4118ab593b5aa9ceea606eddcf190ca7d
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2514231
(cherry picked from commit 1eceeacf73b14e3684457e4e828bd572077e6ad6)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2516965
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>