video: tegra: nvmap: Fix possible memory leak in nvmap_probe
When NVMAP is built as a loadable module, the nvmap_device structure
allocated in nvmap_probe is never freed on failure because the code to
free the memory is only compiled when NVMAP is built into the kernel.
When building NVMAP as a loadable module, only the label 'free_dev'
needs to be compiled out but the call to kfree should remain. However,
rather than fix this by compiling out the label 'free_dev', just use
devm_kzalloc to allocate the memory, because by using this function
memory allocated will be automatically freed on probe failure.

Bug 2879962

Change-Id: I4f860f6858191da8b9390cbc2e2ca8f90a676667
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2514971
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Puneet Saxena <puneets@nvidia.com>
Reviewed-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit