video: tegra: nvmap: Remove declaration of do_set_pte
do_set_pte is declared in nvmap, but it is not being defined anywhere.
We are switching kstable to v5.12-rc8 In this kernel, do_set_pte is
declared in mm.h which leads to conflicting types for do_set_pte.
Hence remove it's declaration from nvmap.

Bug 3294366

Change-Id: I2a9a50451488bc78ae88ec3faa4c37c1641f2575
Signed-off-by: ketanp <ketanp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2517407
Reviewed-by: Puneet Saxena <puneets@nvidia.com>
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit