ALSA: pcm: Don't call sync stop for all DAI links
From Kernel-5.10.27 minor version changes it is observed that audio
suspend test fails with OOM (out of memory) or bad swap prints. These
failures were not seen on Kernel-5.10. Below prints are seen during
system resume and either audio suspend test or kernel warning test or
both are failing frequently in GVS runs. This is more frequently seen
on Jetson TX2 compared to Jetson Xavier.

 1. OOM killer gets invoked by random process. In below case it is
    aplay process.

    [...]
    aplay invoked oom-killer: gfp_mask=0x0(), order=0, oom_score_adj=0
    Out of memory: Killed process 280 (systemd-resolve) total-vm:10584kB,
                   anon-rss:728kB, file-rss:4584kB, shmem-rss:0kB, UID:102
                   pgtables:52kB oom_score_adj:0
    // starts killing few processes
    [...]

 2. Bad swap prints kicked in by random process. In below case it is
    sshd process.

    [...]
    swap_info_get: Bad swap offset entry 2000010b738f
    BUG: Bad page map in process sshd  pte:2000010b738f00 pmd:10e42f003
    addr:0000aaaabda2a000 vm_flags:00000875 anon_vma:0000000000000000
        mapping:ffff0000901a9870 index:a
    file:sshd fault:filemap_fault mmap:nfs_file_mmap readpage:nfs_readpage
    Disabling lock debugging due to kernel taint
    BUG: Bad rss-counter state mm:0000000045fc7086 type:MM_FILEPAGES val:1
    BUG: Bad rss-counter state mm:0000000045fc7086 type:MM_SWAPENTS val:-1
    [...]

Manual bisect points to commit 2c87c1a49c9d ("ALSA: pcm: Assure sync
with the pending stop operation at suspend") and after reverting the same
issue goes away. It moved sync_stop from system resume to suspend and is
called for all DAI links now.

If sync_stop is limited to FE links, then mentioned issues are not seen.
This was the case earlier as well when sync_stop was peformed during system
resume. BE links are internal to AHUB and need not be explicitly handled
in PCM core. Though this change is logical, it is a WAR to the current
issue. Use this to unblock kernel minor version update and further look
to upstream this change and track proper fix.

Bug 200710943
Bug 200710969

Change-Id: Ic3ca99fc56b63e650502de417484bf6dbcf278f7
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2518354
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Manish Bhardwaj <mbhardwaj@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: Uday Gupta <udayg@nvidia.com>
Reviewed-by: Vipin Kumar <vipink@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Manish Bhardwaj <mbhardwaj@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit