platform: tegra: cbb: t23x: don't clear timeout bit
This change adds macro "DONOT_CLR_TIMEDOUT_SLAVE_BIT" to disable
code which clears the timeout bit respective to the slaves who
have timed out. The bit needs to be cleared after resetting the
timedout slave so that the next transactions are send to them.
But currently, there is no support to reset the timed out slaves.
So, disabling it till that feature is added. The code for finding
and printing info about the timed out slaves will still be enabled.

e.g:
Below info about timed out slave will still be printed.
 [  376.188451]    CBB_SN_AXI2APB_8_BLOCK_TMO_STATUS : 0x1

Below info about clearing the reset bit for timed out slave won't be
printed as that code is disabled in this change.
 [  376.189421]    CBB_SN_AXI2APB_8_BLOCK0_TMO : 0x200
 [  376.190313]    Resetting timed-out client 0x200
 [  376.191173]    Readback CBB_SN_AXI2APB_8_BLOCK0_TMO: 0x0

Signed-off-by: Sumit Gupta <sumitg@nvidia.com>
Change-Id: Icce11e8f9f8516ce667ec346061d71586a6bfa79
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2566492
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: Bitan Biswas <bbiswas@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>