misc: nct1008: Fix automatic loading of nct1008
The nct1008 driver is not automatically loaded on boot because the
device-tree MODULE_DEVICE_TABLE definition for the driver is missing.
Add the device-tree MODULE_DEVICE_TABLE for the nct1008 driver so that
the modalias is created and the driver is automatically loaded on boot.

JIRA LS-32
Bug 200721211

Change-Id: I949a1465b12ef22ad1422341f89843ef55bb6e72
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2521821
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: Srikar Srimath Tirumala <srikars@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit