Revert "cpufreq: Update freq_scale upon transition"
This reverts commit c4beaeff963e9cc38cff59e0f2f62d8d7b45c07c.

The change is not required as another change is already present
in upstream kernel as part of a patch which adds a call to
arch_set_freq_scale() at the end of every freq transition in
cpufreq_freq_transition_end(). So, reverting the downstream change.
The related upstream patch is:
 'cpufreq: move invariance setter calls in cpufreq core'

Bug 200674300

Change-Id: I114744e34a2d39761dbdcba1c2dd93e7bcebe6ba
Signed-off-by: Sumit Gupta <sumitg@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2578480
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>