osi: dma: Flush packets on DMA stop
Issue:
When DMA channel is stopped, but packet for that channel is not yet
process from RX queue, it will not be processed and will result in
all other channel packets stuck in the queue.

Fix:
Enable RPF(Rx packet flush) bit when stop DMA and reset in start DMA.
This wil ensure any packet in RX queue for stopped DMA will be flushed
and ensuresmooth working of all other channels.

Bug 200718904

Change-Id: I12f9ecc1754a82d2c2a90d27448348bb4acde349
Signed-off-by: Nagarjuna Kristam <nkristam@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/kernel/nvethernetrm/+/2541985
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Rakesh Goyal <rgoyal@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Bhadram Varka <vbhadram@nvidia.com>
Reviewed-by: Srinivas Ramachandran <srinivasra@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>