core: eqos: Fix DCS setting during filter delete
Issue:
When a L2 filter is deleted, corresponding channel bit is reset
on local variable and OR'ed with register value. Due to OR operation
value was never got reset.

Fix:
Clear register value for all DCS bit and OR with temp value containing
final list of DCS bits.

Bug 200718904

Change-Id: I7c26b06babede74851916ae846a2fe5cf62645f4
Signed-off-by: Nagarjuna Kristam <nkristam@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/kernel/nvethernetrm/+/2541986
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Rakesh Goyal <rgoyal@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Bhadram Varka <vbhadram@nvidia.com>
Reviewed-by: Srinivas Ramachandran <srinivasra@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit