video: tegra: nvmap: put the correct pages in nvmap_ioctl_gup_test
The function nvmap_ioctl_gup_test increments the ref count
for the user-specified pages but it decrements the ref count
for the pages bound to the handle. This could lead to the
handle's pages being freed while still mapped.

This patch only updates the code so that the correct ref count is
updated.

Bug 3324125

Change-Id: Ib05fa255b401db9d95f9395fe3dcbab3049261f2
Signed-off-by: Martin Radev <mradev@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2543036
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: Puneet Saxena <puneets@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit